refactoring went spiffingly. finished it up this morning according to the simplified meta-grammar. added the ability to append meta variables at the beginning of the grammar.
changed the data structures around a bit, added types for Alternatives,
Metas, and Grammars.

grappling with a string lexing problem, as I'd like to be able to use
triple quotes to delimit multiline strings in the meta and in grammars
in general, but for now single quotes will have to do.

time to work on adding actions and deepening the grammar parser, and
then it'll be code generation time.