multiline rule support still doesn't work particularly well, but single line does. going to start writing a code generator now.
shifted the memoizing code into a separate class so that it can be
utilised in generated parsers, though this could also have been done
with inheritance. but i refactored for a separate class file so that's
just how it is going to be for now. maybe i'll go back in the future.