thats much better.
began work on a code generator. going to be a lot of work to transfer
the work already there into something like this, but i am hopeful that
it will go quite quickly. i am still using Van Rossum's work as a guide
but as i head deeper into this my implementation differs more and more
due to the language difference.

i need to consider types, the memoizer, and inheritance much more than
he would have done in python, but i am confident that i can handle it.
little chunks.